+++ instrumentation/logback/logback-appender-1.0/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/logback/appender/v1_0/LogbackSingletons.java
@@ -5,13 +5,11 @@
 
 package io.opentelemetry.javaagent.instrumentation.logback.appender.v1_0;
 
-import static java.util.Collections.emptyList;
-
 import io.opentelemetry.api.GlobalOpenTelemetry;
 import io.opentelemetry.api.incubator.config.DeclarativeConfigProperties;
+import io.opentelemetry.instrumentation.api.config.IncludeExclude;
 import io.opentelemetry.instrumentation.api.incubator.config.internal.DeclarativeConfigUtil;
 import io.opentelemetry.instrumentation.logback.appender.v1_0.internal.LoggingEventMapper;
-import java.util.List;
 
 public class LogbackSingletons {
 
@@ -37,13 +35,13 @@ public class LogbackSingletons {
         config.getBoolean("capture_logstash_marker_attributes/development", false);
     boolean captureLogstashStructuredArguments =
         config.getBoolean("capture_logstash_structured_arguments/development", false);
-    List<String> captureMdcAttributes =
-        config.getScalarList("capture_mdc_attributes/development", String.class, emptyList());
+    IncludeExclude mdcAttributes =
+        LogbackConfig.create(GlobalOpenTelemetry.get()).getMdcAttributes();
 
     mapper =
         LoggingEventMapper.builder()
             .setCaptureExperimentalAttributes(captureExperimentalAttributes)
-            .setCaptureMdcAttributes(captureMdcAttributes)
+            .setMdcAttributes(mdcAttributes)
             .setCaptureCodeAttributes(captureCodeAttributes)
             .setCaptureMarkerAttribute(captureMarkerAttribute)
             .setCaptureKeyValuePairAttributes(captureKeyValuePairAttributes)